fix: split guide champion cost migration

+++ backend/src/main/resources/db/migration/V1__init_schema.sql
@@ -177,7 +177,7 @@ CREATE TABLE tft_guide_champions (
     id BIGINT NOT NULL AUTO_INCREMENT,
     champion_key VARCHAR(100) NOT NULL,
     name VARCHAR(100) NOT NULL,
-    cost INT NOT NULL,
+    cost TINYINT NOT NULL,
     role VARCHAR(50) NOT NULL,
     position VARCHAR(50) NOT NULL,
     image_url VARCHAR(500) NOT NULL,
